requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
Django==4.2.29
django-crontab==0.7.1
django-environ==0.13.0
django-filter==25.1
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
